% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_shiny_gqt_app.R
\name{copy_shiny_gqt_app}
\alias{copy_shiny_gqt_app}
\title{Copy shiny app to user's local server}
\usage{
copy_shiny_gqt_app(copy_to)
}
\arguments{
\item{copy_to}{path to shiny-server app location on user's local server}
}
\value{
Installs webGQT on user's server
}
\description{
Copy shiny app to user's local server
}
\examples{
copy_shiny_gqt_app(copy_to = '/srv/shiny-server/')

}
