% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runwebGQT.R
\name{run_webgqt}
\alias{run_webgqt}
\title{Run webGQT Shiny app}
\usage{
run_webgqt(gqt_path = "/usr/local/gqt/bin/gqt")
}
\arguments{
\item{gqt_path}{Local path for the gqt installation}
}
\value{
Opens a shiny app in the web browser
}
\description{
Run webGQT Shiny app
}
\examples{
run_webgqt(gqt_path = '/usr/local/gqt/bin/gqt')
}
